Indicadores economicos API

+++ Domain/Models/ExchangeRateData.cs
@@ -5,10 +5,10 @@ namespace Domain.Models
     {
         #region Attributes
 
-        public double Purchase { get; set; }
+        public decimal Purchase { get; set; }
         public ExchangeRateResult.ResultError ErrorType { get; set; }
         public int MethodId { get; set; }
-        public double Sell { get; set; }
+        public decimal Sell { get; set; }
         public ExchangeRateResult.ResultType ResultTypeValue { get; set; }
         public bool Success { get; set; }
